refactor(#511): extract advanced orders mixin from alpaca_trading_client.py
- Create AdvancedOrdersMixin in advanced_orders.py with:
  - place_trailing_stop_order
  - place_bracket_order
  - modify_order
  - modify_stop_order
  - cancel_all_orders
- Update AlpacaOrderManager to inherit from mixin
- Reduce alpaca_trading_client.py from 1569 to 1075 lines (31% reduction)
